+++ client-react/src/components/body/msDetails/ShowMSDetails.jsx
@@ -8,6 +8,7 @@ import { Button } from "primereact/button";
 import FormMSDetails from "./FormMSDetails";
 import { useSelector } from "react-redux";
 import { Toolbar } from "primereact/toolbar";
+import { format } from 'date-fns';
 
 const ShowMSDetails = () => {
     const { token } = useSelector((state) => state.token);
@@ -36,16 +37,20 @@ const ShowMSDetails = () => {
     };
 
     const deleteMSDetails = async (rowData) => {
-        if (window.confirm("Are you sure you want to delete this record?")) {
-            {
-                const newDate = new Date(rowData.date)
-                if (newDate.getMonth() === new Date().getMonth()) {
+
+        const newDate = new Date(rowData.date)
+        if (newDate.getMonth() === new Date().getMonth() && newDate.getFullYear() === new Date().getFullYear()) {
+            if (window.confirm("Are you sure you want to delete this record?")) {
+                {
                     const res = await axios.delete(`http://localhost:1111/api/monthlyScholarshipDetails/${rowData._id}`,
                         { headers: { Authorization: `Bearer ${token}` } });
                     getAllMSDetails();
                     setEnableAdd(true)
                 }
             }
+        } else {
+            alert("You can't delete monthly-scholarship-details from last month")
+            return
         }
     };
 
@@ -53,9 +58,13 @@ const ShowMSDetails = () => {
         return (
             <Button label="Update" icon="pi pi-pencil" onClick={() => {
                 const newDate = new Date(rowData.date)
-                if (newDate.getMonth() === new Date().getMonth()) {
-                setMSDetail(rowData);
-                setVisible(true);
+                if (newDate.getMonth() === new Date().getMonth() && newDate.getFullYear() === new Date().getFullYear()) {
+                    setMSDetail(rowData);
+                    setVisible(true);
+                }
+                else {
+                    alert("You can't update monthly-scholarship-details from last month")
+                    return
                 }
             }} ></Button>
         );
@@ -91,6 +100,11 @@ const ShowMSDetails = () => {
             <Button label="Export" icon="pi pi-download" iconPos="right" className="p-button-help" onClick={exportCSV} />
         </React.Fragment>
     );
+    const dateBodyTemplate = (rowData) => {
+        if (rowData.date)
+            return format(rowData.date, 'dd/MM/yyyy')
+        return ""
+    };
 
     return (
         <>
@@ -101,9 +115,8 @@ const ShowMSDetails = () => {
                 <DataTable ref={ms} value={MSDetails} tableStyle={{ minWidth: '50rem' }}>
                     <Column field="sumPerHour" header="SumPerHour"></Column>
                     <Column field="MaximumNumberOfHours" header="MaximumNumberOfHours"></Column>
-                    <Column field="date" header="Date"></Column>
+                    <Column field="date" header="Date" body={dateBodyTemplate}></Column>
                     <Column header="DELETE" body={deleteButton}></Column>
-
                     <Column header="UPDATE" body={updateButton}></Column>
                 </DataTable>
                 {visible && <FormMSDetails visible={visible} setVisible={setVisible} MSDetail={MSDetail} getAllMSDetails={getAllMSDetails} />}
